fix(ml): satisfy acctestconfigdirlint and unconvert for calendar acc tests
- Move inline Config strings to testdata + ConfigDirectory for calendar and
  calendar_event validation/import-wrong-id acceptance tests.
- Drop redundant int64() casts in calendar_event fromAPIModel (unconvert).

+++ internal/elasticsearch/ml/calendar_event/models.go
@@ -148,8 +148,8 @@ func (m *CalendarEventTFModel) fromAPIModel(_ context.Context, apiModel *Calenda
 		endLoc = t.Location()
 	}
 
-	startTime := time.UnixMilli(int64(startMillis)).In(startLoc)
-	endTime := time.UnixMilli(int64(endMillis)).In(endLoc)
+	startTime := time.UnixMilli(startMillis).In(startLoc)
+	endTime := time.UnixMilli(endMillis).In(endLoc)
 
 	m.StartTime = timetypes.NewRFC3339TimeValue(startTime)
 	m.EndTime = timetypes.NewRFC3339TimeValue(endTime)
